requirements.txt
absl-py==2.1.0
altair==5.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
gast==0.5.4
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.22.0
idna==3.6
importlib_metadata==7.0.2
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jws==0.1.3
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==2.0.1
pytz==2024.1
referencing==0.34.0
requests==2.29.0
requests-oauthlib==1.4.0
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.31.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.15.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.18.1
